pyproject.toml
[project]
name = "honuae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromedriver-autoinstaller>=0.6.4",
    "matplotlib>=3.10.3",
    "pip>=25.1.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "setuptools>=80.9.0",
]
